feat:extracting a component for bookings form
